+++ src/app/counter-analytics/day-of-week-histogram/day-of-week-histogram.page.scss
